# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ErrorResponse": {
				"properties": {
					"errorCode": {
						"type": "number",
						"format": "double"
					},
					"errorMessage": {
						"type": "string"
					},
					"errorDetails": {
						"type": "string"
					}
				},
				"required": [
					"errorCode",
					"errorMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"204": {
						"description": "No Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 204,
											"errorMessage": "No Content",
											"errorDetails": "The server successfully processed the request, but there is no content to return"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 401,
											"errorMessage": "Unauthorized",
											"errorDetails": "Authentication credentials are missing or invalid."
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Permission Denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 403,
											"errorMessage": "Permission Denied",
											"errorDetails": "You do not have permission to perform this operation."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 500,
											"errorMessage": "Internal Server Error",
											"errorDetails": "An unexpected error occurred while processing the request"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "perPage",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 400,
											"errorMessage": "Invalid user data",
											"errorDetails": "Please provide valid username, email, and password"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 401,
											"errorMessage": "Unauthorized",
											"errorDetails": "Authentication credentials are missing or invalid."
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Permission Denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 403,
											"errorMessage": "Permission Denied",
											"errorDetails": "You do not have permission to perform this operation."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 500,
											"errorMessage": "Internal Server Error",
											"errorDetails": "An unexpected error occurred while processing the request"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 401,
											"errorMessage": "Unauthorized",
											"errorDetails": "Authentication credentials are missing or invalid."
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Permission Denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 403,
											"errorMessage": "Permission Denied",
											"errorDetails": "You do not have permission to perform this operation."
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 404,
											"errorMessage": "User not found",
											"errorDetails": "The requested user ID was not found in the system"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"errorCode": 500,
											"errorMessage": "Internal Server Error",
											"errorDetails": "An unexpected error occurred while processing the request"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}